更新一处可空值类型生成sql语句错误的问题

+++ Test/User.cs
@@ -12,6 +12,6 @@ namespace Test
         public DateTime Birthday { get; set; }
         public bool Sex { get; set; }
 
-        public int? Value { get; set; }
+        public decimal? Value { get; set; }
     }
 }
